requirements.txt
PyYAML==6.0
requests==2.28.0
PyAudio==0.2.14
typing-extensions==4.8.0
websocket-client==1.3.2